reorganize tests to match botocore
