+++ tests/boto_tests/unit/test_signers.py
