+++ tests/boto_tests/unit/test_utils.py
